File reorg : BaseResult now in its own file
